Enum chronounit::TimeUnit

pub enum TimeUnit {
    Nanoseconds,
    Microseconds,
    Milliseconds,
    Seconds,
    Minutes,
    Hours,
    Days,
}

TimeUnit time unit.

Variants

Nanoseconds

Time unit representing one thousandth of a microsecond.

Microseconds

Time unit representing one thousandth of a millisecond.

Milliseconds

Time unit representing one thousandth of a second.

Seconds

Time unit representing one second.

Minutes

Time unit representing sixty seconds.

Hours

Time unit representing sixty minutes.

Days

Time unit representing twenty-four hours.

Implementations

impl TimeUnit

pub const ZERO: u64 = 0u64

0

pub const ONE: u64 = 1u64

1

pub const THOUSAND: u64 = 1_000u64

1000

pub const SECONDS_PER_MINUTE: u64 = 60u64

60

pub const MINUTES_PER_HOUR: u64 = 60u64

60

pub const HOURS_PER_DAY: u64 = 24u64

24

pub const SECONDS_PER_HOUR: u64 = 3_600u64

pub const SECONDS_PER_DAY: u64 = 86_400u64

pub const NANOS_SCALE: u64 = 1u64

1

pub const THOUSAND_SCALE: u64 = 1_000u64

1000

pub const MICROSECOND_SCALE: u64 = 1_000u64

pub const MILLISECOND_SCALE: u64 = 1_000_000u64

pub const SECOND_SCALE: u64 = 1_000_000_000u64

pub const MINUTE_SCALE: u64 = 60_000_000_000u64

pub const HOUR_SCALE: u64 = 3_600_000_000_000u64

pub const DAY_SCALE: u64 = 86_400_000_000_000u64

pub const NANOS_PER_MICROSECOND: u64 = 1_000u64

pub const NANOS_PER_MILLISECOND: u64 = 1_000_000u64

pub const NANOS_PER_SECOND: u64 = 1_000_000_000u64

pub const NANOS_PER_MINUTE: u64 = 60_000_000_000u64

pub const NANOS_PER_HOUR: u64 = 3_600_000_000_000u64

pub const NANOS_PER_DAY: u64 = 86_400_000_000_000u64

pub const NANOSECONDS_NAME: &'static str = "Nanoseconds"

pub const MICROSECONDS_NAME: &'static str = "Microseconds"

pub const MILLISECONDS_NAME: &'static str = "Milliseconds"

pub const SECONDS_NAME: &'static str = "Seconds"

pub const MINUTES_NAME: &'static str = "Minutes"

pub const HOURS_NAME: &'static str = "Hours"

pub const DAYS_NAME: &'static str = "Days"

pub const NANOSECONDS_LOWER_CASE_NAME: &'static str = "nanoseconds"

pub const MICROSECONDS_LOWER_CASE_NAME: &'static str = "microseconds"

pub const MILLISECONDS_LOWER_CASE_NAME: &'static str = "milliseconds"

pub const SECONDS_LOWER_CASE_NAME: &'static str = "seconds"

pub const MINUTES_LOWER_CASE_NAME: &'static str = "minutes"

pub const HOURS_LOWER_CASE_NAME: &'static str = "hours"

pub const DAYS_LOWER_CASE_NAME: &'static str = "days"

pub fn to_nanos(&self, amount: u64) -> u64

[to_nanos] Converts the given time amount to nanoseconds.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in nanoseconds.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_nanos(1024), 1024);
assert_eq!(TimeUnit::Microseconds.to_nanos(1024), 1024 * 1000);
assert_eq!(TimeUnit::Milliseconds.to_nanos(1024), 1024 * (1000 * 1000));
assert_eq!( TimeUnit::Seconds.to_nanos(1024), 1024 * (1000 * 1000 * 1000));
assert_eq!(TimeUnit::Minutes.to_nanos(1024), 1024 * (1000 * 1000 * 1000 * 60));
assert_eq!(TimeUnit::Hours.to_nanos(1024), 1024 * (1000 * 1000 * 1000 * 60 * 60));
assert_eq!(TimeUnit::Days.to_nanos(1024), 1024 * (1000 * 1000 * 1000 * 60 * 60 * 24));

pub fn to_micros(&self, amount: u64) -> u64

[to_micros] Converts the given time amount to microseconds.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in microseconds.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_micros(1024), 1024 / 1000);
assert_eq!(TimeUnit::Microseconds.to_micros(1024), 1024);
assert_eq!(TimeUnit::Milliseconds.to_micros(1024), 1024 * 1000);
assert_eq!(TimeUnit::Seconds.to_micros(1024), 1024 * (1000 * 1000));
assert_eq!(TimeUnit::Minutes.to_micros(1024), 1024 * (1000 * 1000 * 60));
assert_eq!(TimeUnit::Hours.to_micros(1024), 1024 * (1000 * 1000 * 60 * 60));
assert_eq!( TimeUnit::Days.to_micros(1024), 1024 * (1000 * 1000 * 60 * 60 * 24));

pub fn to_millis(&self, amount: u64) -> u64

[to_millis] Converts the given time amount to milliseconds.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in milliseconds.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_millis(1024), 1024 / 1000 / 1000);
assert_eq!(TimeUnit::Microseconds.to_millis(1024), 1024 / 1000);
assert_eq!(TimeUnit::Milliseconds.to_millis(1024), 1024);
assert_eq!(TimeUnit::Seconds.to_millis(1024), 1024 * 1000);
assert_eq!(TimeUnit::Minutes.to_millis(1024), 1024 * (1000 * 60));
assert_eq!(TimeUnit::Hours.to_millis(1024), 1024 * (1000 * 60 * 60));
assert_eq!(TimeUnit::Days.to_millis(1024), 1024 * (1000 * 60 * 60 * 24));

pub fn to_seconds(&self, amount: u64) -> u64

[to_seconds] Converts the given time amount to seconds.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in seconds.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_seconds(1024), 1024 / 1000 / 1000 / 1000);
assert_eq!(TimeUnit::Microseconds.to_seconds(1024), 1024 / 1000 / 1000);
assert_eq!(TimeUnit::Milliseconds.to_seconds(1024), 1024 / 1000);
assert_eq!(TimeUnit::Seconds.to_seconds(1024), 1024);
assert_eq!(TimeUnit::Minutes.to_seconds(1024), 1024 * 60);
assert_eq!(TimeUnit::Hours.to_seconds(1024), 1024 * (60 * 60));
assert_eq!(TimeUnit::Days.to_seconds(1024), 1024 * (60 * 60 * 24));

pub fn to_minutes(&self, amount: u64) -> u64

[to_minutes] Converts the given time amount to minutes.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in minutes.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_minutes(1024), 1024 / 1000 / 1000 / 1000 / 60);
assert_eq!(TimeUnit::Microseconds.to_minutes(1024), 1024 / 1000 / 1000 / 60);
assert_eq!(TimeUnit::Milliseconds.to_minutes(1024), 1024 / 1000 / 60);
assert_eq!(TimeUnit::Seconds.to_minutes(1024), 1024 / 60);
assert_eq!(TimeUnit::Minutes.to_minutes(1024), 1024);
assert_eq!(TimeUnit::Hours.to_minutes(1024), 1024 * 60);
assert_eq!(TimeUnit::Days.to_minutes(1024), 1024 * (60 * 24));

pub fn to_hours(&self, amount: u64) -> u64

[to_hours] Converts the given time amount to hours.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in hours.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_hours(1024), 1024 / 1000 / 1000 / 1000 / 60 / 60);
assert_eq!(TimeUnit::Microseconds.to_hours(1024), 1024 / 1000 / 1000 / 60 / 60);
assert_eq!(TimeUnit::Milliseconds.to_hours(1024), 1024 / 1000 / 60 / 60);
assert_eq!(TimeUnit::Seconds.to_hours(1024), 1024 / 60 / 60);
assert_eq!(TimeUnit::Minutes.to_hours(1024), 1024 / 60);
assert_eq!(TimeUnit::Hours.to_hours(1024), 1024);
assert_eq!(TimeUnit::Days.to_hours(1024), 1024 * 24);

pub fn to_days(&self, amount: u64) -> u64

[to_days] Converts the given time amount to days.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in days.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_days(1024), 1024 / 1000 / 1000 / 1000 / 60 / 60 / 24);
assert_eq!(TimeUnit::Microseconds.to_days(1024), 1024 / 1000 / 1000 / 60 / 60 / 24);
assert_eq!(TimeUnit::Milliseconds.to_days(1024), 1024 / 1000 / 60 / 60 / 24);
assert_eq!(TimeUnit::Seconds.to_days(1024), 1024 / 60 / 60 / 24);
assert_eq!(TimeUnit::Minutes.to_days(1024), 1024 / 60 / 24);
assert_eq!(TimeUnit::Hours.to_days(1024), 1024 / 24);
assert_eq!(TimeUnit::Days.to_days(1024), 1024);

pub fn to_duration(&self, amount: u64) -> Duration

[to_duration] Converts the given time amount to a std Duration.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in Duration.

Examples

use std::time::Duration;
use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_duration(100), Duration::from_nanos(100));
assert_eq!(TimeUnit::Microseconds.to_duration(100), Duration::from_micros(100));
assert_eq!(TimeUnit::Milliseconds.to_duration(100), Duration::from_millis(100));
assert_eq!(TimeUnit::Seconds.to_duration(1), Duration::from_secs(1));
assert_eq!(TimeUnit::Minutes.to_duration(1), Duration::from_secs(60));
assert_eq!(TimeUnit::Hours.to_duration(1), Duration::from_secs(60 * 60));
assert_eq!(TimeUnit::Days.to_duration(1), Duration::from_secs(60 * 60 * 24));

pub fn to_chrono_duration(&self, amount: u64) -> Duration

[to_chrono_duration] Converts the given time amount to a std Duration.

Arguments

amount - The original time amount, with the unit specified by the caller.

Returns

The converted time amount in Duration.

Examples

use chrono::Duration;
use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.to_chrono_duration(100), Duration::nanoseconds(100));
assert_eq!(TimeUnit::Microseconds.to_chrono_duration(100), Duration::microseconds(100));
assert_eq!(TimeUnit::Milliseconds.to_chrono_duration(100), Duration::milliseconds(100));
assert_eq!(TimeUnit::Seconds.to_chrono_duration(1), Duration::seconds(1));
assert_eq!(TimeUnit::Minutes.to_chrono_duration(1), Duration::minutes(1));
assert_eq!(TimeUnit::Hours.to_chrono_duration(1), Duration::hours(1));
assert_eq!(TimeUnit::Days.to_chrono_duration(1), Duration::days(1));

pub fn value(&self) -> String

[value]

Retrieves the string representation of this TimeUnit.

Returns

A String which represents the formatted value of the object.

Examples

use chronounit::TimeUnit;

assert_eq!(TimeUnit::Nanoseconds.value(), "Nanoseconds");
assert_eq!(TimeUnit::Microseconds.value(), "Microseconds");
assert_eq!(TimeUnit::Milliseconds.value(), "Milliseconds");
assert_eq!(TimeUnit::Seconds.value(), "Seconds");
assert_eq!(TimeUnit::Minutes.value(), "Minutes");
assert_eq!(TimeUnit::Hours.value(), "Hours");
assert_eq!(TimeUnit::Days.value(), "Days");

pub fn sleep(&self, amount: u64)

[sleep]

Sleeps for a specified amount of time according to the TimeUnit,

Arguments

amount - A u64 parameter representing the amount of time to sleep, according to the TimeUnit.

Examples

use std::time::Duration;
use chronounit::TimeUnit;

let start = std::time::Instant::now();
TimeUnit::Milliseconds.sleep(1024);
let duration = start.elapsed();
assert!(duration >= Duration::from_millis(1024));

pub fn closure_sleep<F>(&self, amount: u64, callback: F)
where F: Fn(Duration),

[closure_sleep]

Sleeps for a specified amount of time according to the TimeUnit, then executes a custom sleep function provided by the user/caller.

This function converts the given amount of time to milliseconds based on the TimeUnit enum, constructs a Duration, and then passes it to the provided closure for custom sleep execution.

Arguments

amount - A u64 parameter representing the amount of time to sleep, according to the TimeUnit. callback - A closure that takes a std::time::Duration parameter and performs custom sleep actions.

Examples

use std::thread;
use std::time::Duration;
use chronounit::TimeUnit;
let start = std::time::Instant::now();

TimeUnit::Milliseconds.closure_sleep(1024, |x| {
    assert_eq!(x, Duration::from_millis(1024));
    thread::sleep(x);
});

let duration = start.elapsed();
assert!(duration >= Duration::from_millis(1024));

pub fn closure_chrono_sleep<F>(&self, amount: u64, callback: F)
where F: Fn(Duration),

[closure_chrono_sleep]

Sleeps for a specified amount of time according to the TimeUnit, then executes a custom sleep function provided by the user/caller.

This function converts the given amount of time to milliseconds based on the TimeUnit enum, constructs a Duration, and then passes it to the provided closure for custom sleep execution.

Arguments

amount - A u64 parameter representing the amount of time to sleep, according to the TimeUnit. callback - A closure that takes a chrono::Duration parameter and performs custom sleep actions.

Examples

use std::thread;
use std::time::Duration;
use chronounit::TimeUnit;
let start = std::time::Instant::now();

TimeUnit::Milliseconds.closure_chrono_sleep(1024, |x| {
    assert_eq!(x, chrono::Duration::milliseconds(1024));
    thread::sleep(Duration::from_millis(x.num_milliseconds() as u64));
});

let duration = start.elapsed();
assert!(duration >= Duration::from_millis(1024));

impl TimeUnit

pub fn value_of(value: &str) -> Option<Self>

[value_of]

Returns the corresponding TimeUnit enum based on the provided TimeUnit value/name.

Arguments

value - A string slice representing the value/name of the TimeUnit.

Returns

An Option<Self> containing the matched TimeUnit if a valid value/name is given, otherwise returns None.

use chronounit::TimeUnit;
assert_eq!(TimeUnit::value_of("Nanoseconds"), Some(TimeUnit::Nanoseconds));
assert_eq!(TimeUnit::value_of("Microseconds"), Some(TimeUnit::Microseconds));
assert_eq!(TimeUnit::value_of("Milliseconds"), Some(TimeUnit::Milliseconds));
assert_eq!(TimeUnit::value_of("Seconds"), Some(TimeUnit::Seconds));
assert_eq!(TimeUnit::value_of("Minutes"), Some(TimeUnit::Minutes));
assert_eq!(TimeUnit::value_of("Hours"), Some(TimeUnit::Hours));
assert_eq!(TimeUnit::value_of("Days"), Some(TimeUnit::Days));
assert_eq!(TimeUnit::value_of("Invalid"), None);

pub fn insensitive_case_value_of(value: &str) -> Option<Self>

[insensitive_case_value_of]

Returns the corresponding TimeUnit enum based on the provided TimeUnit value/name. performing a case-insensitive match.

Arguments

value - A string slice representing the value/name of the TimeUnit.

Returns

An Option<Self> containing the matched TimeUnit if a valid value/name is given, otherwise returns None.

use chronounit::TimeUnit;
assert_eq!(TimeUnit::insensitive_case_value_of("Nanoseconds"), Some(TimeUnit::Nanoseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("Microseconds"), Some(TimeUnit::Microseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("Milliseconds"), Some(TimeUnit::Milliseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("Seconds"), Some(TimeUnit::Seconds));
assert_eq!(TimeUnit::insensitive_case_value_of("Minutes"), Some(TimeUnit::Minutes));
assert_eq!(TimeUnit::insensitive_case_value_of("Hours"), Some(TimeUnit::Hours));
assert_eq!(TimeUnit::insensitive_case_value_of("Days"), Some(TimeUnit::Days));
assert_eq!(TimeUnit::insensitive_case_value_of("Invalid"), None);

assert_eq!(TimeUnit::insensitive_case_value_of("NANOSECONDS"), Some(TimeUnit::Nanoseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("MICROSECONDS"), Some(TimeUnit::Microseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("MILLISECONDS"), Some(TimeUnit::Milliseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("SECONDS"), Some(TimeUnit::Seconds));
assert_eq!(TimeUnit::insensitive_case_value_of("MINUTES"), Some(TimeUnit::Minutes));
assert_eq!(TimeUnit::insensitive_case_value_of("HOURS"), Some(TimeUnit::Hours));
assert_eq!(TimeUnit::insensitive_case_value_of("DAYS"), Some(TimeUnit::Days));

assert_eq!(TimeUnit::insensitive_case_value_of("nanoseconds"), Some(TimeUnit::Nanoseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("microseconds"), Some(TimeUnit::Microseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("milliseconds"), Some(TimeUnit::Milliseconds));
assert_eq!(TimeUnit::insensitive_case_value_of("seconds"), Some(TimeUnit::Seconds));
assert_eq!(TimeUnit::insensitive_case_value_of("minutes"), Some(TimeUnit::Minutes));
assert_eq!(TimeUnit::insensitive_case_value_of("hours"), Some(TimeUnit::Hours));
assert_eq!(TimeUnit::insensitive_case_value_of("days"), Some(TimeUnit::Days));