Trait chronounit::formatter::DateTimeFormatter

pub trait DateTimeFormatter {
    // Required methods
    fn of_pattern(&self, pattern: DateTimePattern) -> Box<dyn DateTimeFormatter>;
    fn activated_pattern(&self) -> DateTimePattern;

    // Provided methods
    fn format_date_time_utc_default(&self, datetime: &DateTime<Utc>) -> String { ... }
    fn format_date_time_utc(
        &self,
        datetime: &DateTime<Utc>,
        pattern: DateTimePattern
    ) -> String { ... }
    fn format_naive_date_time_utc_default(
        &self,
        datetime: &NaiveDateTime
    ) -> String { ... }
    fn format_naive_date_time_utc(
        &self,
        datetime: &NaiveDateTime,
        pattern: DateTimePattern
    ) -> String { ... }
    fn format_naive_date_time_default(&self, datetime: &NaiveDateTime) -> String { ... }
    fn format_naive_date_time(
        &self,
        datetime: &NaiveDateTime,
        pattern: DateTimePattern
    ) -> String { ... }
}

DateTimeFormatter date-time formatter

Required Methods

fn of_pattern(&self, pattern: DateTimePattern) -> Box<dyn DateTimeFormatter>

[of_pattern]

Activates the pattern of the formatter.

Examples

use chrono::NaiveDateTime;
use chronounit::formatter::DateTimeFormatter;
use chronounit::formatter::pattern::DateTimePattern;
use crate::chronounit::formatter::DefaultDateTimeFormatter;

let dtf = DefaultDateTimeFormatter::new(DateTimePattern::YyyyMmDd);
let dtf = dtf.of_pattern(DateTimePattern::YyyyMmDdHhMmSs);

assert_eq!(dtf.activated_pattern().value(), DateTimePattern::YyyyMmDdHhMmSs.value());
assert_eq!(dtf.activated_pattern(), DateTimePattern::YyyyMmDdHhMmSs);

fn activated_pattern(&self) -> DateTimePattern

[activated_pattern]

Get the pattern associated with this formatter.

Provided Methods

fn format_date_time_utc_default(&self, datetime: &DateTime<Utc>) -> String

[format_date_time_utc_default]

Formats a UTC date and time according to the formatter default pattern(new/or_pattern).

fn format_date_time_utc( &self, datetime: &DateTime<Utc>, pattern: DateTimePattern ) -> String

[format_date_time_utc]

Formats a UTC date and time according to the specified pattern.

This function takes a reference to a DateTime<Utc> object and a DateTimePattern enum value, then formats the datetime based on the provided pattern, returning a formatted string.

fn format_naive_date_time_utc_default(&self, datetime: &NaiveDateTime) -> String

[format_naive_date_time_utc_default]

Formats a NaiveDateTime according to the formatter default pattern(new/or_pattern).

This function takes a reference to a NaiveDateTime object, then formats the datetime based on the formatter activates pattern, returning a formatted string.

fn format_naive_date_time_utc( &self, datetime: &NaiveDateTime, pattern: DateTimePattern ) -> String

[format_naive_date_time]

Formats a NaiveDateTime according to the specified pattern.

This function takes a reference to a NaiveDateTime object and a DateTimePattern enum value, then formats the datetime based on the provided pattern, returning a formatted string.

fn format_naive_date_time_default(&self, datetime: &NaiveDateTime) -> String

[format_naive_date_time_default]

Formats a NaiveDateTime date and time according to the formatter default pattern(new/or_pattern).

fn format_naive_date_time( &self, datetime: &NaiveDateTime, pattern: DateTimePattern ) -> String

[format_naive_date_time]

Formats a NaiveDateTime according to the specified pattern.

This function takes a reference to a NaiveDateTime object and a DateTimePattern enum value, then formats the datetime based on the provided pattern, returning a formatted string.